First PID loop control added. Uncalibrated

+++ Quadcopter_UniTS/Core/Inc/motor.h
@@ -6,7 +6,7 @@
 #define MOTOR_MAX_PWM     600
 #define MOTOR_MIN_PWM     0
 
-void set_motor_pwm(uint32_t motor_pwm[]);
-void set_motor_pwm_zero(uint32_t motor_pwm[]);
+void set_motor_pwm(uint16_t motor_pwm[]);
+void set_motor_pwm_zero(uint16_t motor_pwm[]);
 
 #endif
